+++ neo.UnitTests/Network/P2P/Payloads/UT_Cosigner.cs
@@ -8,12 +8,12 @@ using Neo.Network.P2P.Payloads;
 namespace Neo.UnitTests.Network.P2P.Payloads
 {
     [TestClass]
-    public class UT_CosignerUsage
+    public class UT_Cosigner
     {
         [TestMethod]
         public void Serialize_Deserialize_Global()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.Global,
                 Account = UInt160.Zero
@@ -22,7 +22,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var hex = "000000000000000000000000000000000000000080";
             attr.ToArray().ToHexString().Should().Be(hex);
 
-            var copy = hex.HexToBytes().AsSerializable<CosignerUsage>();
+            var copy = hex.HexToBytes().AsSerializable<Cosigner>();
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             Assert.AreEqual(attr.Account, copy.Account);
@@ -31,7 +31,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Serialize_Deserialize_CalledByEntry()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.CalledByEntry,
                 Account = UInt160.Zero
@@ -40,7 +40,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var hex = "000000000000000000000000000000000000000001";
             attr.ToArray().ToHexString().Should().Be(hex);
 
-            var copy = hex.HexToBytes().AsSerializable<CosignerUsage>();
+            var copy = hex.HexToBytes().AsSerializable<Cosigner>();
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             Assert.AreEqual(attr.Account, copy.Account);
@@ -49,7 +49,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Serialize_Deserialize_CustomContracts()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.CustomContracts,
                 AllowedContracts = new[] { UInt160.Zero },
@@ -59,7 +59,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var hex = "000000000000000000000000000000000000000010010000000000000000000000000000000000000000";
             attr.ToArray().ToHexString().Should().Be(hex);
 
-            var copy = hex.HexToBytes().AsSerializable<CosignerUsage>();
+            var copy = hex.HexToBytes().AsSerializable<Cosigner>();
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             CollectionAssert.AreEqual(attr.AllowedContracts, copy.AllowedContracts);
@@ -69,7 +69,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Serialize_Deserialize_CustomGroups()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.CustomGroups,
                 AllowedGroups = new[] { ECPoint.Parse("03b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c", ECCurve.Secp256r1) },
@@ -79,7 +79,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var hex = "0000000000000000000000000000000000000000200103b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c";
             attr.ToArray().ToHexString().Should().Be(hex);
 
-            var copy = hex.HexToBytes().AsSerializable<CosignerUsage>();
+            var copy = hex.HexToBytes().AsSerializable<Cosigner>();
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             CollectionAssert.AreEqual(attr.AllowedGroups, copy.AllowedGroups);
@@ -89,7 +89,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Json_Global()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.Global,
                 Account = UInt160.Zero
@@ -98,7 +98,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var json = "{\"account\":\"0x0000000000000000000000000000000000000000\",\"scopes\":\"Global\"}";
             attr.ToJson().ToString().Should().Be(json);
 
-            var copy = CosignerUsage.FromJson(JObject.Parse(json));
+            var copy = Cosigner.FromJson(JObject.Parse(json));
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             Assert.AreEqual(attr.Account, copy.Account);
@@ -107,7 +107,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Json_CalledByEntry()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.CalledByEntry,
                 Account = UInt160.Zero
@@ -116,7 +116,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var json = "{\"account\":\"0x0000000000000000000000000000000000000000\",\"scopes\":\"CalledByEntry\"}";
             attr.ToJson().ToString().Should().Be(json);
 
-            var copy = CosignerUsage.FromJson(JObject.Parse(json));
+            var copy = Cosigner.FromJson(JObject.Parse(json));
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             Assert.AreEqual(attr.Account, copy.Account);
@@ -125,7 +125,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Json_CustomContracts()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.CustomContracts,
                 AllowedContracts = new[] { UInt160.Zero },
@@ -135,7 +135,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var json = "{\"account\":\"0x0000000000000000000000000000000000000000\",\"scopes\":\"CustomContracts\",\"allowedContracts\":[\"0x0000000000000000000000000000000000000000\"]}";
             attr.ToJson().ToString().Should().Be(json);
 
-            var copy = CosignerUsage.FromJson(JObject.Parse(json));
+            var copy = Cosigner.FromJson(JObject.Parse(json));
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             CollectionAssert.AreEqual(attr.AllowedContracts, copy.AllowedContracts);
@@ -145,7 +145,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         [TestMethod]
         public void Json_CustomGroups()
         {
-            var attr = new CosignerUsage()
+            var attr = new Cosigner()
             {
                 Scopes = WitnessScope.CustomGroups,
                 AllowedGroups = new[] { ECPoint.Parse("03b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c", ECCurve.Secp256r1) },
@@ -155,7 +155,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var json = "{\"account\":\"0x0000000000000000000000000000000000000000\",\"scopes\":\"CustomGroups\",\"allowedGroups\":[\"03b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c\"]}";
             attr.ToJson().ToString().Should().Be(json);
 
-            var copy = CosignerUsage.FromJson(JObject.Parse(json));
+            var copy = Cosigner.FromJson(JObject.Parse(json));
 
             Assert.AreEqual(attr.Scopes, copy.Scopes);
             CollectionAssert.AreEqual(attr.AllowedGroups, copy.AllowedGroups);
